✨ Added Host class
